% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-datamods.R, R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{i18n}
\alias{i18n_translations}
\alias{set_i18n}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{datamods}{\code{\link[datamods]{i18n}}, \code{\link[datamods:i18n]{i18n_translations}}, \code{\link[datamods:i18n]{set_i18n}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

